+++ src/routes/ViewAllLogs/ViewAllLogs.test.js
@@ -196,8 +196,7 @@ describe('ViewAllLogs component', () => {
     deleteJobExecutionsSpy.mockClear();
   });
 
-  // TODO: Create separate ticket to fix all the accesibility tests
-  it.skip('should be rendered with no axe errors', async () => {
+  it('should be rendered with no axe errors', async () => {
     const { container } = renderViewAllLogs(defaultQuery);
 
     await runAxeTest({ rootNode: container });