+++ src/routes/tests/Home.test.js
@@ -111,8 +111,7 @@ describe('Home component', () => {
     global.Storage.prototype.getItem.mockReset();
   });
 
-  // TODO: Create separate ticket to fix all the accesibility tests
-  it.skip('should be rendered with no axe errors', async () => {
+  it('should be rendered with no axe errors', async () => {
     const { container } = renderHome(storeWithData);
 
     await runAxeTest({ rootNode: container });