+++ src/settings/MappingProfiles/detailsSections/edit/InstanceDetailsSections/tests/AdministrativeData.test.js
@@ -15,6 +15,7 @@ import {
 
 import { AdministrativeData } from '../AdministrativeData';
 
+import { STATUS_CODES } from '../../../../../../utils';
 import {
   onAdd,
   onRemove,
@@ -66,7 +67,7 @@ describe('Instance "Administrative data" edit component', () => {
   beforeAll(() => {
     global.fetch.mockResolvedValue({
       ok: true,
-      status: 200,
+      status: STATUS_CODES.OK,
       json: async () => ({}),
     });
   });