+++ src/settings/MappingProfiles/detailsSections/view/tests/MappingItemDetails.test.js
@@ -27,8 +27,7 @@ const renderMappingItemDetails = () => {
 };
 
 describe('MappingItemDetails view component', () => {
-  // TODO: Create separate ticket to fix all the accesibility tests
-  it.skip('should be rendered with no axe errors', async () => {
+  it('should be rendered with no axe errors', async () => {
     const { container } = renderMappingItemDetails();
 
     await runAxeTest({ rootNode: container });